:recycle: [refactor] video viewer 위젯 분리 (HH-176)
